requirements.txt
# Библиотека для работы с аннотированными типами в Python, упрощает использование аннотаций типов
annotated-types==0.7.0

# Библиотека для асинхронного программирования, объединяющая функциональность asyncio и trio
anyio==3.6.2

# Библиотека для создания удобных командных интерфейсов (CLI) в Python
click==8.1.7

# Библиотека для работы с цветным выводом в терминале, поддерживает Windows и Unix
colorama==0.4.6

# Современный веб-фреймворк для создания API на Python, основанный на стандартных типах Python
fastapi==0.115.0

# Библиотека для работы с HTTP/1.1, используется для создания HTTP-приложений
h11==0.14.0

# Библиотека для работы с международными доменными именами (IDN), позволяет использовать не-ASCII символы
idna==3.10

# Шаблонизатор для Python, используется для генерации HTML и других текстовых форматов
Jinja2==3.1.4

# Библиотека для безопасного экранирования строк, часто используется вместе с Jinja2
MarkupSafe==2.1.5

# Библиотека для валидации и сериализации данных с использованием аннотаций типов
pydantic==2.9.2

# Основной модуль для Pydantic, обеспечивает базовую функциональность валидации данных
pydantic_core==2.23.4

# Библиотека для определения текущего асинхронного контекста, поддерживает различные асинхронные библиотеки
sniffio==1.3.1

# Минималистичный веб-фреймворк, на котором основан FastAPI, предоставляет инструменты для создания асинхронных приложений
starlette==0.38.6

# Библиотека, предоставляющая дополнительные аннотации типов, которые не входят в стандартную библиотеку Python
typing_extensions==4.12.2

# ASGI-сервер для запуска асинхронных веб-приложений на Python, поддерживает протоколы HTTP/2 и WebSocket
uvicorn==0.30.6

# Библиотека для безопасного создания и проверки токенов, часто используется для работы с сессиями
itsdangerous==2.2.0

# Библиотека для работы с хэшированием паролей, поддерживает множество алгоритмов
passlib[bcrypt]>=1.7.4

# Библиотека для загрузки переменных окружения из файла .env, упрощает конфигурацию приложений
python-dotenv==1.0.1

# Библиотека для обработки multipart/form-data, используется для загрузки файлов через формы
python-multipart==0.0.16

# Библиотека для хэширования паролей с использованием алгоритма bcrypt, обеспечивает безопасность хранения паролей
bcrypt>=4.0.1

# Официальный драйвер MySQL для Python, обеспечивает взаимодействие с базами данных MySQL
mysql-connector-python==9.0.0

# Асинхронный HTTP-клиент для Python, поддерживает HTTP/1.1 и HTTP/2, используется для выполнения HTTP-запросов
httpx==0.27.2